Reorganize and rename front end files
Does not use the atom pattern, just to get a feel for an alternative file layout
