+++ resources/assets/js/Scenes/Home/Filters/LengthFilter.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import Filter from './filter';
-import Button from '../atoms/button';
+import Filter from '../Components/Filter';
+import Button from '../Components/Button';
 
 class LengthFilter extends React.Component {
     renderButtons() {
