+++ resources/assets/js/reducers/filterReducer.js
