+++ resources/assets/js/reducers/lookupReducer.js
