+++ resources/assets/js/reducers/modulesReducer.js
